fix(harness): clear full integration scan failures (issue #2505)
Work-Run: 577ae5cb-41ca-4412-9d0d-dd0c80c1a289
Work-Receipt: g0-r0
