+++ .agents/spec-docs/done/TEST-013-no-global-seam-prevents-a-test-from-being-satisfied-by-host-home-state.md
@@ -83,7 +83,12 @@ this shape, and add the test TC-01 names so the symptom is refused mechanically 
 
 ## User Execution Test Scenarios
 
-Not applicable — no runnable user-facing behaviour changes; verification evidence is recorded in the engineering test plan (TC-01 to TC-03).
+Not applicable.
+
+No runnable user-facing behaviour changes; verification evidence is recorded in the engineering test plan (TC-01 to TC-03).
+
+**Reason:** No runnable user-facing behaviour changes; the engineering test plan records the observable
+proof for the repository-only isolation seam.
 
 Recorded as the rule's required choice rather than skipped.
