+++ scripts/harness/reference-kind-baseline.json
@@ -123,7 +123,6 @@
   ".agents/spec-docs/done/WORKFLOW-004-workflows-build-subcommand.md": 1,
   ".agents/spec-docs/rejected/AGREEMENT-002-complete-august-13-agent-architecture-findings.md": 5,
   ".agents/spec-docs/rejected/ARCH-037-published-contract-hygiene-from-the-arch-audit.md": 1,
-  ".agents/spec-docs/todo/INFRA-162-the-strict-status-check-policy-is-neither-declared-nor-reconciled-so-a-base-fres.md": 2,
   ".agents/specs/harness-composition-design.md": 1,
   ".agents/specs/harness-composition-inventory.md": 4,
   ".agents/tasks/CLI-034-plugin-publish-one-official.md": 1,
@@ -131,7 +130,6 @@
   ".agents/tasks/HARNESS-049-rule-to-skill-agent-refactor.md": 10,
   ".agents/tasks/HARNESS-052-vacuous-green-sweep.md": 1,
   ".agents/tasks/HARNESS-078-single-segment-dotfiles-are-outside-both-name-checks.md": 1,
-  ".agents/tasks/INFRA-162-the-strict-status-check-policy-is-neither-declared-nor-reconciled-so-a-base-fres.md": 1,
   ".agents/tasks/PEER-003-peer-input-coalescing-loses-conversation-turns.md": 1,
   ".agents/tasks/REL-023-changeset-reconciliation.md": 1,
   ".agents/tasks/RUNTIME-002-headless-only-bun-runtime-entry.md": 2,
