+++ scripts/harness/task-lifecycle-legacy-baseline.json
@@ -1,6 +1,6 @@
 {
-  "count": 341,
-  "sha256": "77f1eff21c42111c4a38bc0ae053ad9e80386ac0b49af02b7982ff9031319635",
+  "count": 345,
+  "sha256": "6652618371a510ea76493d4f2e2c8702e9e2f3220f206d3a598261efa5069f3f",
   "owner": "HARNESS-092",
   "note": "Exact pre-canonical archived Task set. No additions or substitutions are allowed."
 }
